AutoHide dock area now always shows pin button independently from DockAreaHasAutoHideButton flag
